+++ src/main/kotlin/compiler/ast/expr/N_BINOP.kt
@@ -96,13 +96,3 @@ class N_OR(left: N_EXPR, right: N_EXPR): N_BINOP("||", left, right, listOf()) {
         setForwardJump("orend")
     }
 }
-
-class N_NULLCOAL(left: N_EXPR, right: N_EXPR): N_BINOP("?:", left, right, listOf()) {
-    override fun code(c: Coder) = with (c.use(this)) {
-        code(left)
-        opcode(O_IFNON)
-        jumpForward("nullskip")
-        code(right)
-        setForwardJump("nullskip")
-    }
-}
