+++ src/main/kotlin/vm/VM.kt
@@ -55,7 +55,9 @@ class VM(
 
     private inline fun push(v: Value) = stack.addFirst(v)
     private inline fun peek() = stack.first()
-    private inline fun pop() = stack.removeFirst().also { if (it.type == Value.Type.VOID) fail(E_VOID, "void in expression") }
+    private inline fun pop() = stack.removeFirst().also {
+        if (it.type == Value.Type.VOID) fail(E_VOID, "void in expression")
+    }
     private inline fun popTwo() = listOf(pop(), pop())
     private inline fun popThree() = listOf(pop(), pop(), pop())
     private inline fun popFour() = listOf(pop(), pop(), pop(), pop())
